Refactored gridsearch.modulefactory to autoencoder module
